Extract Msf::DBManager::Import::Netsparker
MSP-11124
